lift project/directory sync state into global-sync
